pyproject.toml
[project]
name = "C1-Wellbeing-Coach"
version = "0.1.0"
description = "AI-powered wellbeing coaching application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "openai>=1.55.3,<2.0.0",
    "httpx<0.28",
    "python-dotenv>=1.2.2",
    "psycopg[binary]==3.2.13",
    "SQLAlchemy==2.0.34",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]
